Update documentation.
Expands the description of the function to explictly state that the file is initially assumed to be public and checks view-only links and credentials if it is private. A link to the view-only link creation faq is included. The examples have also been updated to point at a demo project for osfr and include demos with both a public file and a private view-only file.